functions/requirements.txt
# Firebase Functions requirements
firebase-functions>=0.1.0
firebase-admin>=6.2.0
google-generativeai>=0.3.0
gspread>=6.0.0
oauth2client>=4.1.3
requests>=2.31.0
python-dotenv>=1.0.0
